requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9
torch==2.3.0
torchvision==0.18.0
opencv-python-headless==4.9.0.80
Pillow==10.3.0
scikit-image==0.23.2
numpy==1.26.4
scipy==1.13.0
pydantic==2.7.1
joblib==1.4.2
scikit-learn==1.5.0
